Added pi calculation
-strengthened compile flags
-added release option to Makefile
-fixed all warning caused by -Wall

+++ src/Tester.cpp
@@ -25,7 +25,7 @@ bool Tester::runTests(void) {
 					 "before running.\n";
 		return true;
 	}
-	for (int i = 0; i < tests.size(); i++) {
+	for (size_t i = 0; i < tests.size(); i++) {
 		Test test = tests[i];
 		bool success = false;
 		if (!test.isExceptionTest)
